+++ App.js
@@ -6,12 +6,14 @@ import { DrawerContent } from "./screens/DrawerContent";
 import SupportScreen from "./screens/SupportScreen";
 import SettingScreen from "./screens/SettingScreen";
 import BookmarkScreen from "./screens/BookmarkScreen";
+import RootStackScreen from "./screens/RootStackScreen";
 
 const Drawer = createDrawerNavigator();
 
 const App = () => {
   return (
     <NavigationContainer>
+      {/*<RootStackScreen />*/}
       <Drawer.Navigator drawerContent={(props) => <DrawerContent {...props} />}>
         <Drawer.Screen name="HomeDrawer" component={MainTabScreen} />
         <Drawer.Screen name="SupportScreen" component={SupportScreen} />
