+++ screens/RootStackScreen.js
@@ -1,0 +1,17 @@
+import React from "react";
+
+import { createStackNavigator } from "@react-navigation/stack";
+
+import LogInScreen from "./LogInScreen";
+import SignUpScreen from "./SignUpScreen";
+
+const RootStack = createStackNavigator();
+
+const RootStackScreen = ({ navigation }) => (
+  <RootStack.Navigator headerMode="none">
+    <RootStack.Screen name="LogInScreen" component={LogInScreen} />
+    <RootStack.Screen name="SignUpScreen" component={SignUpScreen} />
+  </RootStack.Navigator>
+);
+
+export default RootStackScreen;
